% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello!}
\usage{
hello()
}
\value{
printout
}
\description{
{\preformatted{#\if{latex,text}{\n\\deqn{\n\\frac{\\Sigma(|positive|)-\\Sigma(|negative|)}{\\Sigma(|positive|)+\\Sigma(|negative|)}x100\n}{\n\\frac{\\Sigma(|positive|)-\\Sigma(|negative|)}{\\Sigma(|positive|)+\\Sigma(|negative|)}x100\n}}"
## attr(,"class")
## [1] "Rdtext"
}\preformatted{## [1] "\\WMCXoBctmlJ1LA\if{latex,text}{\n\\deqn{\nf(x)= {\\frac{1}{\\sigma\\sqrt{2\\pi}}}e^{- {\\frac {1}{2}} (\\frac {x-\\mu}{\\sigma})^2}\n}{\nf(x)= {\\frac{1}{\\sigma\\sqrt{2\\pi}}}e^{- {\\frac {1}{2}} (\\frac {x-\\mu}{\\sigma})^2}\n}}"
## attr(,"class")
## [1] "Rdtext"
}<!-- ```{r, include=rd, results='asis'} -->
<!-- eval(parse(text=sprintf("\\\\Sexpr[results=rd, stage=build]{katex::math_to_rd('\%s')}", equation))) -->
<!-- ``` -->


}
This works with built in example
\Sexpr[results=rd, stage=build]{
katex::math_to_rd(katex::example_math())
}
}
\examples{
hello()
}
